% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_audit.R
\name{cols_audit}
\alias{cols_audit}
\title{Audit Data Worksheet column names}
\usage{
cols_audit()
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the Audit Data worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
}
