% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_equipment.R
\name{cols_equipment}
\alias{cols_equipment}
\title{Equipment Worksheet column names}
\usage{
cols_equipment()
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the Equipment worksheet in Oregon DEQ's continuous data submission template xlsx file v3.0.
}
