% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_mloc.R
\name{cols_mloc}
\alias{cols_mloc}
\title{Monitoring Location worksheet column names}
\usage{
cols_mloc()
}
\value{
vector of column names.
}
\description{
Retrieve column names for the Monitoring Locations worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
Includes internal staff columns: "Reachcode", "Measure", "LLID", "River Mile", "Permanent Identifier", "Monitoring Location Status ID", and
"Monitoring Location Status Comment".
}
