% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_prepost.R
\name{cols_prepost}
\alias{cols_prepost}
\title{PrePost Worksheet column names}
\usage{
cols_prepost()
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the PrePost worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
}
