% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_projects.R
\name{cols_projects}
\alias{cols_projects}
\title{Projects Worksheet column names}
\usage{
cols_projects()
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the Projects worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
}
