% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_results.R
\name{cols_results}
\alias{cols_results}
\title{Results Worksheet column names}
\usage{
cols_results()
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the Results worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
Includes internal staff columns: "accDQL", "precDQL", and "rDQL.
}
