% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conqc.R
\docType{data}
\name{conqc}
\alias{conqc}
\title{Continuous QC data quality matrix}
\format{
An object of class \code{data.frame} with 19 rows and 8 columns.
}
\usage{
data(conqc)
}
\description{
Summary of minimum values defining data quality levels
}
\details{
\itemize{
\item Characteristic.Name:	Parameter name.
\item Qccalc: Statistical calculation.
\item prec_A: Minimum value defining data quality level A precision.
\item prec_B: Minimum value defining data quality level B precision.
\item acc_A: Minimum value defining data quality level A accuracy.
\item acc_B:	Minimum value defining data quality level B accuracy.
\item Source: Oregon DEQ source QAPP document.
}
}
\examples{
conqc <- odeqcdr::conqc

}
\keyword{datasets}
