% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contin_export.R
\name{contin_export}
\alias{contin_export}
\title{Export continuous result information from to template xlsx file.}
\usage{
contin_export(file, org, projects, mloc, deployment, results, prepost, audits)
}
\arguments{
\item{file}{The path and file name to the output xlsx file.}

\item{org}{data frame holding Organization Details}

\item{projects}{data frame with Projects}

\item{mloc}{data frame holding Monitoring_Locations}

\item{deployment}{data frame holding Deployment}

\item{results}{data frame holding Results}

\item{prepost}{data frame holding PrePost}

\item{audits}{data frame holding Audit_Data}
}
\description{
Export monitoring results into Oregon DEQ's continuous data submission template xlsx file format v2.03.
This function will export a list with each element holding a data frame of the information for each spreadsheet in the template.
The cell colors and quality control elements from the actual template are not included in the output xlsx.
}
\details{
This function relies upon the \link{openxlsx} package.
}
\seealso{
\link{openxlsx}
}
