% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contin_results_csv.R
\name{contin_results_csv}
\alias{contin_results_csv}
\title{Import continuous result information from a template csv file.}
\usage{
contin_results_csv(path = NULL, col_rename = NULL)
}
\arguments{
\item{path}{The path to the directory where the csv files are saved. Default is NULL which implements the \link{choose.dir} function on windows OS
or \link[rstudioapi:file-dialogs]{rstudioapi::selectDirectory} on Mac/Linux OS when using Rstudio. If Rstudio is not avaibile a path must be provided.}

\item{col_rename}{named character vector, with new names as values, and old names as names.c("new_name=old_name")}
}
\value{
data frame
}
\description{
Retrieve monitoring results from Oregon DEQ's continuous data submission template v2.03 that has been saved as a csv
This function will read all csv files from a selected directory, combine them all using \link{rbind} and return them as a single dataframe.
The column names must be the same in every csv file.
}
\details{
Column names are made into syntactically valid names acceptable by R.
}
