% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dql_dos.R
\name{DOsat_DQLs}
\alias{DOsat_DQLs}
\title{Assigns DO concentration DQLs to DO saturation}
\usage{
DOsat_DQLs(df = results)
}
\arguments{
\item{results}{Dataframe of results after manual grade corrections have been applied to DO concentration values}
}
\value{
Dataframe of all results with final grade.
}
\description{
This function takes the accDQL, precDQL, and rDQL from DO concentration results and applies them to DO saturation
results, as the measurements are taken from the same sensor. This is consistent with how the lab treats DOsaturation
DQLs. This cript should be run after DO concentration results have been assigned final grades.
}
