% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dql_accuracy.R
\name{dql_accuracy}
\alias{dql_accuracy}
\title{Assign accuracy Data Quality Levels (DQLs) to continuous data}
\usage{
dql_accuracy(prepost, results, prepost_only = FALSE)
}
\arguments{
\item{prepost}{Data frame of the prepost data generated using odeqcdr::import_contin().}

\item{results}{Data frame of the results data generated using odeqcdr::import_contin().}

\item{prepost_only}{Boolean to indicate if the prepost data frame should be returned with a new columns for the accuracy DQL,
and absolute difference between the result and references values. Default is FALSE.}
}
\value{
Vector of the precision DQL indexed in the same order as the result input. Or if audit_only=TRUE a data frame.
}
\description{
This function assigns an accuracy data quality level to parameters measured in the field using pre- and post-deployment checks.
The data quality levels are assigned following DEQ’s Data Quality Matrix (DEQ, 2013).
The data quality matrix defines the accuracy and precision criteria for equipment calibration and field audits respectively respectively.
}
\details{
use odeqcdr::dql_precision() to assign the data quality level using field duplicates, field audits, or split samples.

Inputs into this function must retrieved from Oregon DEQ's continuous data submission template xlsx file v2.03.
Use odeqcdr::import_contin() to get the list that hold the data frames used for this function's inputs.

Oregon Department of Environmental Quality (DEQ). 2013. "Data validation criteria for water quality parameters measured in the field.
DEQ04-LAB-0003-QAG Version5.0." http://www.oregon.gov/deq/FilterDocs/DataQualMatrix.pdf
}
