% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dql_precision.R
\name{dql_precision}
\alias{dql_precision}
\title{Assign precision Data Quality Levels (DQLs) to continuous data}
\usage{
dql_precision(audits, results, deployment, audits_only = FALSE)
}
\arguments{
\item{audits}{Data frame of the audit data generated using \code{\link[=contin_import]{contin_import()}}.}

\item{results}{Data frame of the results data generated using \code{\link[=contin_import]{contin_import()}}.}

\item{deployment}{Data frame of the deployment data generated using \code{\link[=contin_import]{contin_import()}}.}

\item{audit_only}{Boolean to indicate if the audit data frame should be returned with new columns for the audit sample DQL,
corresponding result value, result units, absolute difference between the result and audit, and an auto assigned row number. Default is FALSE.}
}
\value{
Vector of the precision DQL indexed in the same order as the result input. Or if audit_only=TRUE a data frame.
}
\description{
This function assigns a precision data quality level to parameters measured in the field using replicate measurements
such as field duplicates, field audits, or split samples. The data quality levels are assigned following DEQ’s Data Quality Matrix (DEQ, 2013).
The data quality matrix defines the accuracy and precision criteria for equipment calibration and field audits respectively respectively.
}
\details{
use \link{dql_accuracy(}) to assign the data quality level using calibration verifications and/or pre- and post-deployment checks.

Inputs into this function must retrieved from Oregon DEQ's continuous data submission template xlsx file v2.03.
Use \code{\link[=contin_import]{contin_import()}} to get the list that hold the data frames used for this function's inputs.

Oregon Department of Environmental Quality (DEQ). 2013. "Data validation criteria for water quality parameters measured in the field.
DEQ04-LAB-0003-QAG Version5.0." http://www.oregon.gov/deq/FilterDocs/DataQualMatrix.pdf
}
