% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_check.R
\name{dst_check}
\alias{dst_check}
\title{Check that date and time conform to changes between Daylight Time and Standard Time.}
\usage{
dst_check(
  df,
  mloc_col = "Monitoring.Location.ID",
  char_col = "Characteristic.Name",
  equip_col = "Equipment.ID",
  date_col = "Activity.Start.Date",
  time_col = "Activity.Start.Time",
  tz_col = "Activity.Start.End.Time.Zone"
)
}
\arguments{
\item{df}{Data frame to be evaluated.}

\item{mloc_col}{Column name in df holding the Monitoring Location ID. Default is "Monitoring.Location.ID".}

\item{char_col}{Column name in df holding the Characteristic Name. Default is "Characteristic.Name".}

\item{equip_col}{Column name in df holding the Equipment ID. Default is "Equipment.ID"}

\item{date_col}{Column name in df holding the POSIXct date values. Default is "Activity.Start.Date"}

\item{time_col}{Column name in df holding the POSIXct time values. Default is "Activity.Start.Time".}

\item{tz_col}{Column name in df holding the timezone string values. Default is tz_col ="Activity.Start.End.Time.Zone".}
}
\value{
Vector of corrected datetime in POSISXct.
}
\description{
This function will evaluate if an hour was added or subtracted correctly to account
for the switch from daylight time to standard time and vice versa. The check can only
be used on timeseries occurring on or after the start of standard time on October 27, 1985 at 2:01 AM.
}
\details{
The dst check fails and time is adjusted when either of the two conditions occur:
\enumerate{
\item When there are results recorded between 2:00 AM and 2:59 AM when daylight savings begins.
As an example, in 2020 the switch from standard time to daylight time (+1 hour) occurred
on March 8, 2020 at 2:00 AM. In a time series where samples are collected
every 30 minutes there should be no results recorded from 2:00 AM until 2:59 AM.
If there are results recorded during this period it’s likely an adjustment
to daylight savings time was not made.
\item When there is only one result recorded for each datetime between 1:00 AM and 1:59 AM
after the transition back to standard time (-1 hour). As an example, in 2020 the switch from daylight savings
back to standard time occurred on November 1, 2020 at 2:00 AM resulting in a repeat of time
between 1:00 AM and 1:59 AM. If there are not duplicate times recorded it is likely
the adjustment to standard time was not made.
}

When one of the conditions are detected in a timeseries, the time at the beginning
of the deployment period is considered the correct time and the appropriate adjustment
for daylight savings is made from that point forward. The dst check evaluates
each deployment [Monitoring.Location.ID, Characteristic.Name, Equipment.ID]
and any adjustments are made only to an individual deployment. A manual review of the results
is recommend.

Daylight Savings in Oregon since 1986

1986 - 2006:
Daylight savings time began by adding on hour (+1) on the first Sunday in April at 2:00 AM.
Daylight savings time ended by subtracting an hour (-1) on the last Sunday in October at 2:00 AM.

2007 - Current:
Daylight savings time begins by adding an hour (+1) on the second Sunday in March at 2:00 AM.
Daylight savings time ends by subtracting an hour (-1) on the first Sunday in November at 2:00 AM.
}
