% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_combine.R
\name{dt_combine}
\alias{dt_combine}
\title{Combine PoSIXct date, time, and timezone columns into a single POSIXct datetime.}
\usage{
dt_combine(
  df,
  date_col = "Activity.Start.Date",
  time_col = "Activity.Start.Time",
  time_val = NULL,
  tz_col = "Activity.Start.End.Time.Zone"
)
}
\arguments{
\item{df}{Data frame to be evaluated.}

\item{date_col}{Column name in df holding the POSIXct date values. Default is "Activity.Start.Date"}

\item{time_col}{Column name in df holding the POSIXct time values. Default is "Activity.Start.Time". If NULL uses date_col or time_val if provided.}

\item{time_val}{string in format 'hh:mm:ss' representing the time to use for all rows. Default is NULL. If NULL uses date_col or time_col if provided.}

\item{tz_col}{Column name in df holding the timezone string values. Acceptable values
include "PST", "PDT", "MST", "MDT" or Olson Names such as 'America/Los_Angeles' for Pacific Time
zone or 'America/Boise' for Mountain time zone.
See \code{\link[base]{OlsonNames}} for more info. Default is tz_col="Activity.Start.End.Time.Zone".}
}
\value{
vector of POSIXct datetime
}
\description{
This function will combine POSIXct date and time values pulled from an excel using \code{\link{contin_import}} to make a POSIXct datetime.
There are various options on how to handle the time and timezone column. If the datetime falls in a daylight savings break NA is returned.
}
