% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_shift.R
\name{dt_shift}
\alias{dt_shift}
\title{Check for datetime shifts (Temperature only)}
\usage{
dt_shift(results, return_df = FALSE)
}
\arguments{
\item{results}{Data frame of the results data generated using \code{\link{contin_import}}.}

\item{return_df}{TRUE means the results data frame will be returned with the dt_shift column. FALSE only returns the dt_shift vector. Default is FALSE.}
}
\value{
data frame.
}
\description{
A datetine shift is evaluated by checking that the daily maximum water temperature
occurs between 13:00 and 19:00. If the daily maximum falls outside this hour range the function
flags all temperature results on that day as an anomaly where dt_shift=TRUE. A datetime shift may
indicate an issue with the time not being adjusted to the correct time zone
(i.e. still in UTC/GMT), a copy/paste/transcription error, or invalid results. results dataframe must
have a column called 'datetime' formatted as POSIXct.
}
