% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Get measure value along Oregon DEQ's NHDH_OR_931v220.}
\usage{
get_measure(pid, x, y)
}
\arguments{
\item{pid}{The NHD Permanent_Identifier}

\item{x}{The longitude in decimal degrees}

\item{y}{The latitude in decimal degrees}
}
\value{
sf object with data frame columns for Permanent_Identifier, ReachCode, measure, and snap_distance
}
\description{
The function will query Oregon DEQ's NHD feature service to determine the measure value along a target NHD flowline. The Permanent Identifier (pid) is
used to select a specific reach from the feature service. The x and y coordinates are snapped to the nearest vertex on the selected flowline and the measure value determined. The point is
This function is intended to be a helper function for \link{launch_map} although it may be used independently also.
}
\details{
The feature service can be accessed at https://arcgis.deq.state.or.us/arcgis/rest/services/WQ/NHDH_ORDEQ/MapServer/1
}
