% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_contin.R
\name{import_contin}
\alias{import_contin}
\title{Import continuous result information from a template xlsx file.}
\usage{
import_contin(file)
}
\arguments{
\item{file}{The path and file name to template xlsx file.}
}
\value{
list of each continuous template data
}
\description{
Retrieve monitoring results from Oregon DEQ's continuous data submission template xlsx file v2.03.
This function will read the template and return a list with each list element holding a dataframe of the information for each spreadsheet.
Any rows with all NAs are removed.
}
