% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inchars.R
\name{inchars}
\alias{inchars}
\title{Replace Invalid Characters.}
\usage{
inchars(x)
}
\arguments{
\item{x}{string or vector}
}
\value{
vector without invalid characters.
}
\description{
This function will replace invalid characters.
}
\details{
'@' symbol is replaced with 'at'
Any of the following: ` ~ ! # $ \% ^ & * \ | ; : ' " < > / ? are replaced with an underscore '_'.
Parentheses and brackets are replaced with an underscore.
Spaces are replaced with and underscore.
}
