% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_map.R
\name{launch_map}
\alias{launch_map}
\title{Launch a web map in a browser to review monitoring stations}
\usage{
launch_map(mloc)
}
\arguments{
\item{mloc}{Data frame of the monitoring location data generated using \code{\link[=contin_import]{contin_import()}}.}
}
\value{
Launches a leaflet map within a Shiny app
}
\description{
Launch a web map in a browser to review monitoring stations
}
