% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch Shiny app for further review}
\usage{
launch_shiny()
}
\value{
Launches a Shiny app
}
\description{
A shiny application to review data
}
\details{
The data must first be formatted as a named list and saved as an RData file.

The object must be named 'shiny_list' with the named listed elements as follows:
'Deployment'
'Audit_Stats'
'Results_Anom'

Example code:

shiny_list <-list(Deployment=df1.deployment,
Audit_Stats=df3.audits.dql,
Results_Anom=df5.results.anom)

save(shiny_list, file=paste0(CDR_template_for_shiny_review.Rdata"))
}
