% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats2.R
\name{sumstats2}
\alias{sumstats2}
\title{Generate AWQMS Summary Statistics}
\usage{
sumstats2(results, deployment, project_id)
}
\arguments{
\item{results}{Data frame of the results data generated using \code{\link{contin_import}} or \code{\link{contin_volmon_import}}.}

\item{deployment}{Data frame of the deployment data generated using \code{\link{contin_import}} or \code{\link{contin_volmon_import}}.}

\item{project_id}{Unique Project ID from the projects data frame generated using \code{\link{contin_import}} or \code{\link{contin_volmon_import}}. Only accepts one project ID.}
}
\value{
returned data frame.
}
\description{
Calculates summary statistics but does not determine DQLs. Returns a data frame
formatted for AWQMS upload. Version 0.12, commit 8de029f2565fe4645a06e6eaee36f545557888db
was the last instance where this function was the original sumstats.
}
\details{
Summary statistics calculated include:
\itemize{
\item Temperature, water: Daily Maximum, Daily Minimum, Daily Mean, and 7-day mean of the daily maximums (7DMADMax);
\item Dissolved oxygen (DO): Daily Maximum, Daily Minimum, Daily Mean, 7-day mean of the daily minimums (7DMADMin), 7-day mean of the daily mean (7DMADMean), and 30-day mean of the daily mean (30DMADMean);
\item All others: Daily Maximum, Daily Minimum, and Daily Mean
}

Rolling means are trailing means reported on the last day of the 7-day or 30-day period.
}
