% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats_DQL.R
\name{sumstats_DQL}
\alias{sumstats_DQL}
\title{Generate AWQMS Summary Stats (with DQLs) Upload File}
\usage{
sumstats_DQL(results, deployment, project_id)
}
\arguments{
\item{results}{Data frame of the results data generated using \code{\link[=contin_volmon_import]{contin_volmon_import()}}.}

\item{deployment}{Data frame of the deployment data generated using \code{\link[=contin_volmon_import]{contin_volmon_import()}}.}

\item{project_id}{Unique Project ID from the projects data frame generated using  \code{\link[=contin_volmon_import]{contin_volmon_import()}}. Only accepts one project ID.}
}
\value{
returned data frame.
}
\description{
Generate AWQMS Summary Stats (with DQLs) Upload File
}
