% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstr.R
\name{tstr}
\alias{tstr}
\title{Returns a string of index values cooresponding to each vector element that is TRUE}
\usage{
tstr(x)
}
\arguments{
\item{x}{Boolean vector. NA values are treated the same as a FALSE.}
}
\value{
string
}
\description{
This works the same as \link{which} except each index value is collapsed into a single
string separated by a comma and space. +1 is added to the index to equal
the row number in excel assuming the header row is the first row.
If all values are FALSE, NA is returned. NA values in x are treated the same as a FALSE.
}
\seealso{
\link{valid_values} \link{valid_values_check}
}
