% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_deploy.R
\name{update_deploy}
\alias{update_deploy}
\title{Update the deployment Worksheet to match a different template version}
\usage{
update_deploy(deploy, from = 2, to = 3)
}
\arguments{
\item{deploy}{The deployment dataframe.}

\item{from}{Template version that 'deploy' is being updated from. Numeric value equal to 2 or 3.}

\item{to}{Template version that 'deploy' will be updated to. Numeric value equal to 2 or 3.}
}
\value{
updated deployment dataframe.
}
\description{
Reorganize the deployment columns to match a specific xlsx template configuration.
}
