% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dql_update.R
\name{update_DQL}
\alias{update_DQL}
\title{Manual update of rDQL field}
\usage{
update_DQL(df, rows, DQL, comment = NA_character_)
}
\arguments{
\item{df}{Dataframe to modify. Should be df4.results. No need to specify if part or pipe structure}

\item{rows}{Rows to modify. Get these from reviewing the shinyapp}

\item{DQL}{Updated final DQL value}

\item{comment}{Comment to be used in Result.Comment field}
}
\value{
Dataframe with updated DQL and comment
}
\description{
This function is to provide an ability to manually update rDQL field after manual review in shiny app for the volmon
program. Input is intended to be the df4.results dataframe must contain the following fields: rDQL, row.results,
Result.Comment
}
