% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values.R
\name{valid_values}
\alias{valid_values}
\title{Return a vector of valid values for a column.}
\usage{
valid_values(col = NULL)
}
\arguments{
\item{col}{Name of the column in the continuous data submission template xlsx file v2.03 to return valid values.
The name can be formatted as shown the xlsx template or in R format derived using make.names(). if col=NULL All valid values are
returned as a list with the column name as the list lookup. Default is NULL}
}
\value{
vector of valid values or list of all valid values.
}
\description{
This function will return valid values for certain columns in the continuous data submission template xlsx file v2.03. NA is included as a valid value for columns that are optional or conditional.
}
