% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_deploy.R
\name{cols_deploy}
\alias{cols_deploy}
\title{Deployment Worksheet column names}
\usage{
cols_deploy(ver = 2)
}
\arguments{
\item{ver}{numeric version of the template. Valid values are ver=2 (the default) or ver=3. The column names are different between the two versions.}
}
\value{
Vector of column names.
}
\description{
Retrieve column names for the Deployment worksheet in Oregon DEQ's continuous data submission template xlsx file v2.03.
}
