% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_check.R
\name{dst_check}
\alias{dst_check}
\title{Check that date and time conform to changes between Daylight Time and Standard Time.}
\usage{
dst_check(
  df,
  mloc_col = "Monitoring.Location.ID",
  char_col = "Characteristic.Name",
  equip_col = "Equipment.ID",
  date_col = "Activity.Start.Date",
  time_col = "Activity.Start.Time",
  tz_col = "Activity.Start.End.Time.Zone"
)
}
\arguments{
\item{df}{Data frame to be evaluated.}

\item{mloc_col}{Column name in df holding the Monitoring Location ID. Default is "Monitoring.Location.ID".}

\item{char_col}{Column name in df holding the Characteristic Name. Default is "Characteristic.Name".}

\item{equip_col}{Column name in df holding the Equipment ID. Default is "Equipment.ID"}

\item{date_col}{Column name in df holding the POSIXct date values. Default is "Activity.Start.Date"}

\item{time_col}{Column name in df holding the POSIXct time values. Default is "Activity.Start.Time".}

\item{tz_col}{Column name in df holding the timezone string values. Default is tz_col ="Activity.Start.End.Time.Zone".}
}
\value{
Vector of corrected datetime in POSISXct.
}
\description{
This function will evaluate if an hour was added or subtracted correctly to account
for the switch from daylight time to standard time and vice versa.
For example in 2020, the switch from standard time to daylight time occurred
on March 8, 2020 at 2:00 AM. In a time series where samples are collected
every 30 minutes there should be no results recorded from 2:00 AM until 2:59 AM
on March 8. If there are results recorded during this period it’s likely an adjustment
for daylight savings was not made. This function attempts to find these occurrences and
adds or subtracts an hour to all datetime values during the entire daylight savings
or standard time period. A manual review of the results is recommend.
}
