% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_map.R
\name{launch_map}
\alias{launch_map}
\title{Launch a web map in a browser to review and edit monitoring location information}
\usage{
launch_map(mloc)
}
\arguments{
\item{mloc}{Data frame of the monitoring location data generated using \code{\link[=contin_import]{contin_import()}}.}
}
\value{
Launches a leaflet map within a Shiny app. Returns mloc data frame with any saved changes on app close.
}
\description{
Launch a web map in a browser to review and edit monitoring location information
}
