% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeqcdr-package.R
\docType{package}
\name{odeqcdr-package}
\alias{odeqcdr}
\alias{odeqcdr-package}
\title{odeqcdr: ODEQ Continuous Data Review}
\description{
This package provides QA/QC support for review of continuous data entered into the Oregon Department of Environmental Quality's continuous water quality data templates.
}
\author{
\strong{Maintainer}: Ryan Michie \email{Ryan.Michie@deq.oregon.gov}

Authors:
\itemize{
  \item Travis Pritchard \email{Travis.Pritchard@deq.oregon.gov}
}

}
\keyword{internal}
