% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values_check.R
\name{valid_values_check}
\alias{valid_values_check}
\title{Return a vector of boolean values indicating if the value is a valid value}
\usage{
valid_values_check(col, vals)
}
\arguments{
\item{col}{Name of the column in the continuous data submission template xlsx file v2.03 to return valid values.
The name can be formatted as shown the xlsx template or in R format derived using \code{\link[=make.names]{make.names()}}.}

\item{val}{vector of values to evaluate.}
}
\value{
vector of valid values or list of all valid values.
}
\description{
This function will return valid values for certain columns in the continuous
data submission template xlsx file v2.03. NA is included as a valid value for columns that are optional or conditional.
}
